nmfc-api-backend/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
python-dotenv
pydantic
